// modules/BSIM/src/types.tsx
export enum CoinTypes {
  CONFLUX = 'CONFLUX',

  ETHEREUM = 'ETHEREUM',
  CALLISTO = 'CALLISTO',
  GOCHAIN = 'GOCHAIN',
  ETHEREUM_CLASSIC = 'ETHEREUM_CLASSIC',
  POA = 'POA',
  VECHAIN = 'VECHAIN',
  WANCHAIN = 'WANCHAIN',
  TRON = 'TRON',
  ICON = 'ICON',
  TOMO = 'TOMO',
  BITCOIN = 'BITCOIN',
  LITECOIN = 'LITECOIN',
  BITCOINCASH = 'BITCOINCASH',
  DASH = 'DASH',
  ZCOIN = 'ZCOIN',
  ZCASH = 'ZCASH',
  BINANCE = 'BINANCE',
  RIPPLE = 'RIPPLE',
  TEZOS = 'TEZOS',
  STELLAR = 'STELLAR',
  KIN = 'KIN',
  AION = 'AION',
  NIMIQ = 'NIMIQ',
  THUNDERTOKEN = 'THUNDERTOKEN',
  ATOM = 'ATOM',
  KAVA = 'KAVA',
  DOGECOIN = 'DOGECOIN',
  THETA = 'THETA',
  ONTOLOGY = 'ONTOLOGY',
  GROESTL = 'GROESTL',
  VIACOIN = 'VIACOIN',
  QTUM = 'QTUM',
  ZELCASH = 'ZELCASH',
  ZILLIQA = 'ZILLIQA',
  IOTEX = 'IOTEX',
  RAVEN = 'RAVEN',
  WAVES = 'WAVES',
  AETERNITY = 'AETERNITY',
  NEBULAS = 'NEBULAS',
  FIO = 'FIO',
  DECRED = 'DECRED',
  ALGORAND = 'ALGORAND',
  NANO = 'NANO',
  DIGIBYTE = 'DIGIBYTE',
  HARMONY = 'HARMONY',
  NEAR = 'NEAR',
  SOLANA = 'SOLANA',
  ELROND = 'ELROND',
  POLKADOT = 'POLKADOT',
  SMARTCHAIN = 'SMARTCHAIN',
  SMARTCHAINLEGACY = 'SMARTCHAINLEGACY',
  FILECOIN = 'FILECOIN',
}